Remove 'user_type' field, add role-switching function and user role specification
The 'user_type' field was removed from the database schema as it was redundant with the 'role' field. A function was added in JdbcUserDao to allow switching a user's role. The Register.vue file was modified to allow the user to specify their role (DJ or regular user) during the registration process. Additionally, a new testdata.sql file was created for testing purposes.