Add new functionalities to the dashboard and user management
This update introduces several changes to the dashboard and user management features. It adds additional functionalities such as retrieving all parties, displaying parties, getting users in a party, getting a playlist in a party, handling features of a created song, and handling genres of created songs.

This is done by creating new services in Vue for the party, playlist, song, and user. New mappings and methods are also added in the java DAO interface and classes. A new component for party display is introduced in Vue. Also, UI changes are made to the home screen, and security issue is fixed where certain pages required authentication.

These changes are aimed at enhancing user experience and functionality of the app.