***EMERGENCY COMMIT JDBCPARTYDAO METHOD GETUSERSINPARTY WAS RETURNING A LIST<STRING> INSTEAD OF LIST<USER> WHEN GETTING ALL USERS IN A PARTY***
Remove JbdcPlaylistDaoTests.java and update Vue and Java files

The JbdcPlaylistDaoTests.java file is deleted as the tests have been refactored and moved to a different test class. Significant UI updates are made to the Vue components "PartyContainer.vue", "PartyDisplay.vue", "Home.vue" and "App.vue", improving responsiveness, design and adding a section to display users in a party.

The "PartyController.java" now returns a list of users in a party instead of a single user. New comments added in the "JdbcUserDao.java" for clarity. Exception handling improved in "getUsersInParty" method of "PartyController.java".

Major changes in "testdata.sql" to generate more varied test data.

These changes enhance usability and clarity of the codebase and provide more robust testing.